Reduce circular dependencies: MediaRenderer <-> ModelViewer
